now using shared DataManagerViewModel
