+++ HangTab/ViewModels/SettingsViewModel.cs
@@ -1,19 +1,16 @@
 ﻿using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
-using CommunityToolkit.Mvvm.Messaging;
 
 using HangTab.Enums;
-using HangTab.Messages;
 using HangTab.Services;
 using HangTab.ViewModels.Base;
 
 namespace HangTab.ViewModels;
 [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage(Justification = "This is a ViewModel for the UI and does not require unit tests.")]
 public partial class SettingsViewModel(
-    IDatabaseService databaseService,
-    IDialogService dialogService,
     ISettingsService settingsService,
-    IThemeService themeService) : ViewModelBase
+    IThemeService themeService,
+    DataManagerViewModel dataManager) : ViewModelBase
 {
     [ObservableProperty]
     private int _totalSeasonWeeks;
@@ -39,53 +36,15 @@ public partial class SettingsViewModel(
     public override async Task LoadAsync() => await Loading(InitializeSettingsAsync);
 
     [RelayCommand]
-    private async Task DeleteAllDataAsync()
-    {
-        if (!await dialogService.Ask("Delete", "Are you sure you want to delete ALL data?", "Yes", "No"))
-        {
-            return;
-        }
-
-        if (await databaseService.DeleteAllData())
-        {
-            SendSystemResetMessage();
-            await dialogService.ToastAsync("All data has been deleted");
-        }
-        else
-        {
-            await dialogService.AlertAsync("Critical Error", "Error occurred while deleting the databases!", "Ok");
-        }
-    }
+    private async Task DeleteAllDataAsync() => await dataManager.DeleteAllDataAsync();
 
     [RelayCommand]
-    private async Task StartNewSeasonAsync()
-    {
-        if (!await dialogService.Ask("Season Reset", "Are you ready to start a new season and reset all bowler hangings?", "Yes", "No"))
-        {
-            return;
-        }
-
-        if (await databaseService.DeleteSeasonData())
-        {
-            SendSystemResetMessage();
-            await dialogService.ToastAsync("A new season has started");
-        }
-        else
-        {
-            await dialogService.AlertAsync("Critical Error", "An error occurred while starting a new season!", "Ok");
-        }
-    }
+    private async Task StartNewSeasonAsync() => await dataManager.StartNewSeasonAsync();
 
     private Task InitializeSettingsAsync()
     {
         TotalSeasonWeeks = settingsService.TotalSeasonWeeks;
         DarkThemeEnabled = settingsService.Theme == (int)Theme.Dark;
         return Task.CompletedTask;
-    }
-    
-    private void SendSystemResetMessage()
-    {
-        settingsService.CurrentWeekPrimaryKey = 0;
-        WeakReferenceMessenger.Default.Send(new SystemResetMessage());
     }
 }
